backend/requirements.txt
fastapi
uvicorn
strawberry-graphql
sqlalchemy
psycopg2-binary
